Refaktoroi: kuntailmoituslomake yleiskäyttöisemmäksi;
Refaktoroi lomake toimimaan myös ilman hakutilannelistalta peräisin
olevia tietoja, jotta sitä voi käyttää myös oppijan detaljinäkymässä.
