+++ valpas-web/src/state/apitypes/henkilo.ts
@@ -1,17 +1,15 @@
 import { ISODate, Oid } from "../common"
 
-export type HenkilöLaajatTiedot = {
+export interface HenkilöTiedot {
   oid: Oid
-  hetu?: string
   syntymäaika?: ISODate
   etunimet: string
   sukunimi: string
+}
+
+export type HenkilöLaajatTiedot = HenkilöTiedot & {
+  hetu?: string
   turvakielto: boolean
 }
 
-export type HenkilöSuppeatTiedot = {
-  oid: Oid
-  syntymäaika?: ISODate
-  etunimet: string
-  sukunimi: string
-}
+export type HenkilöSuppeatTiedot = HenkilöTiedot
