+++ valpas-web/src/views/ilmoituslomake/IlmoitusForm.less
@@ -1,5 +1,5 @@
-@import "../../../style/colors.less";
-@import "../../../style/typography.less";
+@import "../../style/colors.less";
+@import "../../style/typography.less";
 
 @ilmoitusform-frame-color: @color-blue;
 @ilmoitusform-padding: 1.5rem;
